+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 
-## **3.0.0** - *2020-xx-xx*
+## **3.0.1** - *2020-04-30*
+* example deploy shell scripts added;
+* example [pm2](https://www.npmjs.com/package/pm2) configs added;
+* `Facade/Gallery` - go to prev/next photo with swipe;
+* `TODO.md` added;
+
+## **3.0.0** - *2020-04-19*
 * dynamic content blocks with Quill WYSIWYG editor in `Dashboard`;
 * `Dashboard`:
     * app toolbar made responsive;